server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.8.0
certifi==2024.8.30
python-dotenv==1.0.1
werkzeug==3.0.3
google-auth==2.40.3
PyJWT==2.10.1
requests==2.32.3
Flask-Mail==0.10.0
